+++ docker/mongodb-enterprise-tests/tests/authentication/sharded_cluster_scram_sha_1_connectivity.py
@@ -9,7 +9,7 @@ from tests.authentication.sha1_connectivity_tests import SHA1ConnectivityTests
 class TestShardedClusterSHA1Connectivity(SHA1ConnectivityTests):
     @fixture
     def yaml_file(self):
-        return "sharded-cluster-scram-sha-1.yaml"
+        return "sharded-cluster-explicit-scram-sha-1.yaml"
 
     @fixture
     def mdb_resource_name(self):